Add TODOs, remove unused params, rename test and file.
